Double default value of EventTimerBUfferSize to 10k

+++ runners/samza/src/main/java/org/apache/beam/runners/samza/SamzaPipelineOptions.java
@@ -77,7 +77,7 @@ public interface SamzaPipelineOptions extends PipelineOptions {
   void setSystemBufferSize(int consumerBufferSize);
 
   @Description("The maximum number of event-time timers to buffer in memory for a PTransform")
-  @Default.Integer(5000)
+  @Default.Integer(10000)
   int getEventTimerBufferSize();
 
   void setEventTimerBufferSize(int eventTimerBufferSize);
